\name{bliss}
\alias{bliss}
\title{bliss method for Qualitative assay data analysis}
\usage{
bliss(AT,sdose,sm,sr,tdose,tm,tr)
}
\description{
A convenient way to analyse biaassay data using bliss method.
}
\arguments{
  \item{AT}{Estimated potency.}
  \item{sdose}{Dose of standard group in the experiment.}
  \item{sm}{Sample number of standard group in the experiment.}
  \item{sr}{Reaction number of standard group in the experiment.}
  \item{tdose}{Dose of trial group in the experiment.}
  \item{tm}{Sample number of trial group in the experiment.}
  \item{tr}{Reaction number of trial group in the experiment.}
}
