\name{direct_assay}
\alias{direct_assay}
\title{direct assay method}
\usage{
direct_assay(AT,ys,nt,yt)
}
\description{
A convenient way to analyse biaassay data using direct_assay method.
}
\arguments{
  \item{AT}{Estimated potency.}
  \item{ys}{Response values of standard group in the experiment.}
  \item{nt}{Sample number of each trial group in the experiment.}
  \item{yt}{Response values of all trial groups in the experiment.}
}
