\name{logistic_4p}
\alias{logistic_4p}
\title{logistic_4p method for Bioassay data analysis}
\usage{
logistic_4p(AT,dose,rep,ys,yt)
}
\description{
A convenient way to analyse biaassay data using logistic method.
}
\arguments{
  \item{AT}{Estimated potency.}
  \item{dose}{Dose of Trial/Standard group in the experiment.}
  \item{rep}{replication number for each dose.}
  \item{ys}{Response values of standard group in the experiment.}
  \item{yt}{Response values of trial group in the experiment.}
}
